enterprise-use-cases/data-analysis/requirements.txt
bedrock-agentcore
strands-agents
boto3
pandas
